backward compatibility with pykeepass 3.0.2: no attachment
This is required because PyPI does not allow installing modules from
URLS for security reasons. https://www.python.org/dev/peps/pep-0508/
is not supported in PyPI

If it was, it would be possible to add a dependency to a fork of
pykeepass.

See https://github.com/pschmitt/pykeepass/issues/137 for a discussion